Turn on/off use of nocc for lexing editor content
This is part-way to a self-contained version of the lexer for AVR assembly
sources.  Calling nocc to tokenise the source file and parsing the resulting
output is workable if process launch/shutdown costs are relatively small.
On Linux this is true;  on Windows with cygwin/mingw, less true (and that
results in a pretty sluggish editor).  Controlled with USE_NOCC_LEXER in
avrasmlexer.h.